+++ src/App.tsx
@@ -1,7 +1,8 @@
 import { useEffect, useState } from 'react';
-import Grid from "./components/Grid";
-import GridControls from "./components/GridControls";
-import Game from './game/Grid';
+import Grid from "./app/components/Grid";
+import GridControls from "./app/components/GridControls";
+import { getGenerationSpeed } from './app/util';
+import Game from './class/Grid';
 
 let game: Game = {} as Game;
 let intervalID: NodeJS.Timeout = {} as NodeJS.Timeout;
@@ -9,7 +10,8 @@ let intervalID: NodeJS.Timeout = {} as NodeJS.Timeout;
 function App() {
   const [boardNeedsInitiaization, setBoardInitialization] = useState(true);
   const [generation, setGeneration] = useState(0);
-  const [generationSpeed, setGenerationSpeed] = useState(1000);
+  const [generationSpeed, setGenerationSpeed] = useState(1);
+  const [isGameRunning, setIsGameRunning] = useState(false);
 
   useEffect(() => {
     if (boardNeedsInitiaization) {
@@ -25,15 +27,21 @@ function App() {
     }
   });
 
-  function runGame() {
+  function runGameInterval() {
     intervalID = setInterval(() => {
       game.next();
       setGeneration(game.getGenerations())
-    }, generationSpeed)
+    }, getGenerationSpeed(generationSpeed))
+  }
+
+  function runGame() {
+    runGameInterval();
+    setIsGameRunning(true);
   }
 
   function stopGame() {
     clearInterval(intervalID)
+    setIsGameRunning(false);
   }
 
   function nextGeneration(e: React.MouseEvent<HTMLButtonElement, MouseEvent>) {
@@ -44,6 +52,11 @@ function App() {
 
   function updateGenerationSpeed(value: number) {
     setGenerationSpeed(value)
+
+    if (isGameRunning) {
+      clearInterval(intervalID)
+      runGameInterval();
+    }
   }
 
   function resetBoard() {
