+++ src/app/components/Grid.tsx
@@ -1,10 +1,14 @@
-import Game from "../game/Grid";
+import Game from "../../class/Grid";
 
 interface Props {
   game: Game;
 }
 
 function Grid({ game }: Props) {
+  if (!game) {
+    return null;
+  }
+
   const gameJSON = game.getStatus && game.getStatus();
   const gridJSON = JSON.stringify(gameJSON);
