+++ src/core/coordinates.spec.ts
