+++ src/core/game.spec.ts
