+++ src/core/tests/stillLifes.spec.ts
@@ -1,4 +1,4 @@
-import gameOfLife from "../../game";
+import gameOfLife from "../game";
 
 describe('still lifes', () => {
   describe('block', () => {
