+++ src/interfaces/index.ts
@@ -1,0 +1,4 @@
+// The Life Grid state is represented as a Hasg
+export interface LifeGrid {
+  [coordinate: string]: boolean;
+}
